# openapi.yaml
openapi: 3.0.3
info:
  title: Membership Service API
  description: |
    OAuth 2.0 API for authenticating users and retrieving member information.

    This API implements Resource Owner Password Flow with client credentials validation,
    allowing external services to authenticate users securely and access their full profile data.

    For detailed integration guide with code examples, see API_INTEGRATION.md
  version: 1.0.0
  contact:
    name: Membership Service Support
    email: admin@example.com
servers:
- url: https://your-membership-service.com/api
  description: Production server
- url: http://localhost:8787/api
  description: Local development server
security:
- BearerAuth: []
...
paths:
  /oauth/token:
    post:
      tags:
      - OAuth
      summary: Obtain access token
      description: |
        OAuth 2.0 token endpoint implementing Resource Owner Password Flow.

        External services must provide both their own credentials (client_id, client_secret)
        and the user's credentials (username, password) to obtain an access token.

        The access token is valid for 1 hour and should be cached per user session.
      operationId: getToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - grant_type
              - client_id
              - client_secret
              - username
              - password
              properties:
                grant_type:
                  type: string
                  enum: [ password ]
                  description: OAuth grant type (must be "password")
                  example: password
                client_id:
                  type: string
                  description: Your service's OAuth client ID
                  example: client_abc123def456789
                client_secret:
                  type: string
                  format: password
                  description: Your service's OAuth client secret
                  example: secret_xyz789abc123def456
                username:
                  type: string
                  description: User's email address or membership number
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: User's password
                  example: user_secure_password
                scope:
                  type: string
                  description: Requested scope (optional)
                  example: user_info
                  default: user_info
      responses:
        '200':
          description: Access token successfully issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad request (invalid parameters or unsupported grant type)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  value:
                    error: invalid_request
                    error_description: username and password are required
                unsupportedGrantType:
                  value:
                    error: unsupported_grant_type
                    error_description: Only password grant type is supported
        '401':
          description: Authentication failed (invalid credentials)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidClient:
                  value:
                    error: invalid_client
                    error_description: Invalid client credentials
                invalidGrant:
                  value:
                    error: invalid_grant
                    error_description: Invalid user credentials
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth/user-info:
    get:
      tags:
      - OAuth
      summary: Get user profile information
      description: |
        Retrieve complete user profile data using a valid access token.

        Returns all available user information including personal details,
        contact information, educational background, and membership status.
      operationId: getUserInfo
      security:
      - BearerAuth: []
      responses:
        '200':
          description: User profile successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Unauthorized (missing, invalid, or expired token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingToken:
                  value:
                    error: invalid_token
                    error_description: Missing or invalid Authorization header
                expiredToken:
                  value:
                    error: invalid_token
                    error_description: Token is invalid or expired
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid_token
                error_description: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth/introspect:
    post:
      tags:
      - OAuth
      summary: Introspect access token
      description: |
        Check if an access token is valid and retrieve its metadata.

        Useful for debugging and token validation. Requires client authentication.
      operationId: introspectToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              - client_id
              - client_secret
              properties:
                token:
                  type: string
                  description: Access token to introspect
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                client_id:
                  type: string
                  description: Your service's OAuth client ID
                  example: client_abc123def456789
                client_secret:
                  type: string
                  format: password
                  description: Your service's OAuth client secret
                  example: secret_xyz789abc123def456
      responses:
        '200':
          description: Token introspection result
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/IntrospectionResponseActive'
                - $ref: '#/components/schemas/IntrospectionResponseInactive'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /oauth/token endpoint
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [ Bearer ]
          description: Token type (always "Bearer")
          example: Bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        scope:
          type: string
          description: Granted scope
          example: user_info
    UserInfoResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        membership_number:
          type: string
          description: Unique membership identifier
          example: MEM123456
        ar_name:
          type: string
          description: Arabic name
          example: اسم العضو
        latin_name:
          type: string
          description: Latin/English name
          example: Member Name
        whatsapp:
          type: string
          description: WhatsApp number
          example: '+1234567890'
        email:
          type: string
          format: email
          description: Email address
          example: user@example.com
        sex:
          type: string
          description: Gender
          example: Male
        birth_date:
          type: string
          format: date
          description: Date of birth
          example: '1990-01-01'
        country:
          type: string
          description: Country
          example: Saudi Arabia
        city:
          type: string
          description: City
          example: Riyadh
        district:
          type: string
          description: District/neighborhood
          example: Al Olaya
        university:
          type: string
          description: University name
          example: King Saud University
        major:
          type: string
          description: Field of study
          example: Computer Science
        graduation_year:
          type: string
          description: Graduation year
          example: '2024'
        blood_type:
          type: string
          description: Blood type
          example: O+
        phone:
          type: string
          description: Phone number
          example: '+1234567890'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: invalid_grant
        error_description:
          type: string
          description: Human-readable error description
          example: Invalid user credentials
    IntrospectionResponseActive:
      type: object
      properties:
        active:
          type: boolean
          example: true
        client_id:
          type: string
          example: client_abc123def456789
        username:
          type: string
          example: MEM123456
        scope:
          type: string
          example: user_info
        exp:
          type: integer
          description: Token expiration timestamp
          example: 1738886400
        iat:
          type: integer
          description: Token issued at timestamp
          example: 1738882800
    IntrospectionResponseInactive:
      type: object
      properties:
        active:
          type: boolean
          example: false
...